Rename resource to account rather than user
